+++ utils/api.py
@@ -31,7 +31,6 @@ from typing import (
     TypeVar,
     cast,
 )
-from urllib.parse import urljoin
 
 import datarobot as dr
 import instructor
@@ -122,8 +121,8 @@ def initialize_deployment() -> tuple[RESTClientObject, str]:
     try:
         dr_client = dr.Client()
         chat_agent_deployment_id = LLMDeployment().id
-        deployment_chat_base_url = urljoin(
-            dr_client.endpoint, f"/deployments/{chat_agent_deployment_id}/"
+        deployment_chat_base_url = (
+            f"{dr_client.endpoint.rstrip('/')}/deployments/{chat_agent_deployment_id}/"
         )
         return dr_client, deployment_chat_base_url
     except ValidationError as e: