+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Data cleansing now automatically removes leading and trailing whitespace from string columns
 - Fixed chat endpoint when DATAROBOT_ENDPOINT has a trailing slash
+- Fixed the raw data preview for SAP (react)
 
 ### Changed